+++ bot_setup.py
@@ -1,0 +1,4 @@
+from aiogram import Bot
+from config import BOT_TOKEN
+
+bot = Bot(token=BOT_TOKEN)
