+++ config.py
@@ -1,0 +1,4 @@
+import token_for_bot
+BOT_TOKEN = token_for_bot.bot_token
+NUM_PLAYERS = 2
+NUM_ROLLS = 3
