+++ db/database.py
@@ -1,0 +1,5 @@
+from db.game_db import SQLiteDB
+
+db = SQLiteDB("game.db")
+db.connect()
+db.init_game_table()
