+++ gameplay/states.py
@@ -1,0 +1,6 @@
+from aiogram.dispatcher.filters.state import State, StatesGroup
+
+
+class GameStates(StatesGroup):
+    play = State()
+    rolls = State()
